pyproject.toml
[tool.poetry]
name = "everything2prompt"
version = "0.1.0"
description = "Turn any data source into a prompt."
authors = ["Bilal Chughtai <brchughtaii@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
pyyaml = "^6.0.2"
pydantic = "^2.11.7"
jinja2 = "^3.1.6"
todoist-api-python = "^3.1.0"
dotenv = "^0.9.9"
icalendar = "^6.3.1"
requests = "^2.31.0"


[tool.poetry.group.dev.dependencies]
ipykernel = "^6.30.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
